flask_app_DL/requirements.txt
Flask
onnxruntime
numpy
Pillow
Werkzeug
gunicorn